py/djsite/requirements.txt
django-rq==0.8.0
python-social-auth==0.2.13
requests==2.8.1
oauthlib==1.0.3
Django==1.8.5
django-redis-cache==1.6.3
google-api-python-client==1.4.2
gunicorn
MySQL-python
PyCrypto
